requirements.txt
streamlit
pandas
scikit-learn
joblib
openai
python-dotenv
